refactor(assets): extract useReceiveForm from ReceiveForm.vue (#185)
Move the ReceiveForm script verbatim into a co-located useReceiveForm
composable so the SFC is a thin projection, per the house Vue style.
Behaviour-preserving: the inline currency-select handler becomes a named
onSelectCurrency, and the selected-currency option now reads the existing
`currency` computed instead of re-indexing assets in the template.

Receive keeps its own composable (not shared with Send): its getRoute
source/sink order is the reverse of Send's, and its balance-fetch path
differs — kept per-form so the asymmetry stays visible.
